refactor: remove hardcode type and derive typeId from master data
